+++ DAOS/daoAudio.py
@@ -11,6 +11,7 @@
 #########################################################################################
 
 import redis
+import Configuracion.constantesPrefijosClaves as constantesPrefijosClaves
 
 
 #########################################################################################
