+++ DAOS/daoGlobal.py
@@ -1,0 +1,2 @@
+import redis
+import Configuracion.constantesPrefijosClaves as constantesPrefijosClaves
